requirements.txt
anyio==3.6.1
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
cftime==1.6.2
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cookiecutter==2.1.1
dask==2022.9.2
fastapi==0.85.0
fsspec==2022.8.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
locket==1.0.0
MarkupSafe==2.1.1
netCDF4==1.6.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
partd==1.3.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.12.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
text-unidecode==1.3
toolz==0.12.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
xarray==2022.9.0

